perf(types): decouple useComponentProps from the component-types barrel
