+++ src/runtime/components/Theme.vue
@@ -2,7 +2,7 @@
 import { type VNode, computed } from 'vue'
 import defu from 'defu'
 import { injectThemeContext, provideThemeContext } from '../composables/useComponentProps'
-import type { ThemeContextDefaults, ThemeDefaults, ThemeUI } from '../composables/useComponentProps'
+import type { ThemeContextDefaults, ThemeDefaults, ThemeUI } from '../types/theme'
 
 export interface ThemeProps {
   /**
