+++ src/runtime/composables/index.ts
@@ -3,6 +3,7 @@ export * from './defineShortcuts'
 export * from './useAvatarGroup'
 export * from './useComponentIcons'
 export * from './useComponentProps'
+export type { ThemeDefaults, ThemeUI, ThemeContextDefaults } from '../types/theme'
 export * from './useContentSearch'
 export * from './useEditorMenu'
 export * from './useFieldGroup'
